requirements.txt
alembic==0.9.6
bcrypt==3.1.4
cffi==1.11.2
click==6.7
coverage==4.4.1
Flask==0.12.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.2
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.1
pycparser==2.18
PyJWT==1.5.0
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.0
Werkzeug==0.14
